app, droplet resources: rework validation for environment variables
ValidateFuncs nested in `Elem` don't actually appear to check against map
keys. additionally, go's regexp doesn't support negation, so it's
inappropriate to use for validation here. replace with basic function
replicating the intent.

also add validation ensuring empty environment variables aren't set. this
is because internally to the cf client they collide with the signalling
for "unsetting" an env var in an update call, meaning it can't successfully
persist an empty env var to app state and this will upset terraform when
it realizes it's not getting back the app state it thought it set.